# Add RasterSummary Operation