--- a/scripts/benchmark
+++ b/scripts/benchmark
@@ -23,6 +23,7 @@ huc10_file = 'RasterGroupedCount_LowerSchuylkillRiver_HUC10.json'
 huc12_file = 'RasterGroupedCount_LittleNeshaminy_HUC12.json'
 rga_file = 'RasterGroupedAverage.json'
 rlj_file = 'RasterLinesJoin_Schuylkill_HUC08.json'
+rs_file = 'RasterSummary_Schuylkill_HUC08.json'
 
 def make_rga_request(input_file, operation_name):
     print("Timing {} ->\n".format(operation_name))
@@ -64,6 +65,9 @@ def time_raster_grouped_average():
 def time_raster_lines_join():
     make_rga_request(rlj_file, 'HUC8 RasterLinesJoin')
 
+def time_raster_summary():
+    make_rga_request(rs_file, 'HUC8 RasterSummary')
+
 def time_huc12():
     make_rgc_request(huc12_file, 'HUC12 RasterGroupedCount')
 
@@ -74,7 +78,9 @@ def time_huc8():
     make_rgc_request(huc8_file, 'HUC8 RasterGroupedCount')
 
 parser = argparse.ArgumentParser(description='Test Geoprocessing service \
-    response times for RasterGroupedCount and RasterGroupedAverage operations.')
+    response times for RasterGroupedCount, RasterGroupedAverage, and \
+    RasterSummary operations.')
+
 parser.add_argument('--huc8', help='Time a HUC8 with 1-3 layers',
                     action='store_true')
 parser.add_argument('--huc10', help='Time a HUC10 with 1-3 layers',
@@ -85,12 +91,15 @@ parser.add_argument('--rga', help='Time a RasterGroupedAverage operation',
                     action='store_true')
 parser.add_argument('--rlj', help='Time a RasterLinesJoin operation',
                     action='store_true')
+parser.add_argument('--rs', help='Time a RasterSummary operation',
+                    action='store_true')
 
 args = vars(parser.parse_args(sys.argv[1:]))
 
 if True not in args.values():
     time_raster_lines_join()
     time_raster_grouped_average()
+    time_raster_summary()
     time_huc12()
     time_huc10()
     time_huc8()
@@ -100,3 +109,4 @@ else:
     if args['huc12']: time_huc12()
     if args['rga']: time_raster_grouped_average()
     if args['rlj']: time_raster_lines_join()
+    if args['rs']: time_raster_summary()