… examples/MapshedJob_RasterSummary.json
… scripts/benchmarking_input_data/RasterSummary_Schuylkill_HUC08.json