@kellyi 2018-01-02
```sh
$ ./scripts/benchmark --rs
Timing HUC8 RasterSummary ->

HUC8 RasterSummary, run 1 -> 9.404321 s
HUC8 RasterSummary, run 2 -> 5.998905 s
HUC8 RasterSummary, run 3 -> 6.975526 s
HUC8 RasterSummary, run 4 -> 5.116356 s
HUC8 RasterSummary, run 5 -> 4.945487 s

HUC8 RasterSummary average -> 6.488119 s
```