@kellyi 2018-01-02
```sh
$ http :8090/run < examples/MapshedJob_RasterSummary.json
HTTP/1.1 200 OK
Content-Length: 124
Content-Type: application/json
Date: Tue, 02 Jan 2018 16:12:40 GMT
Server: akka-http/10.0.9

{
    "result": [
        {
            "avg": 799.9116994958924,
            "max": 3430.0,
            "min": -38.0
        },
        {
            "avg": 1.2253132788318364,
            "max": 28.67453956604004,
            "min": 0.0
        }
    ]
}
```