@kellyi 2018-01-02
Working well! Going to take a look at the code now.

I'd be interested in very quickly doing https://github.com/WikiWatershed/mmw-geoprocessing/issues/56 since the current README doesn't have accurate instructions wrt setting up this project for development. I believe we can dump the `Usage` section altogether and replace it with:

* Requirements
* Getting Started
* Development
* Scripts
* Ports

...each of which should be pretty small, mostly just enumerating things