@rajadain 2018-01-02
I'd prefer to do #56 on its own since this is already sharing points with https://github.com/WikiWatershed/model-my-watershed/issues/2346, but I agree its been left for quite some time and should be done soon.