@rajadain 2018-01-02
Thanks for taking a look! Going to squash the fixups and merge shortly.